feat(Text): add LinkCta component
