+++ catalog/pages/links/index.js
@@ -3,13 +3,14 @@ import { ThemeProvider } from "styled-components";
 
 import { Container, Row, Column } from "../../../src/components/Grid";
 import Spacing from "../../../src/components/Spacing";
-import { Link, Text } from "../../../src/components/Text";
+import { Link, LinkCta, Text } from "../../../src/components/Text";
 
 export default {
   path: "/links",
   title: "Links",
   imports: {
     Link,
+    LinkCta,
     Text,
     ThemeProvider,
     Container,
