+++ src/components/Text/Link.js
@@ -1,53 +1,34 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
+import Base from "./Base.styles";
 import * as PT from "./PropTypes";
-import { typography } from "../../theme";
-import { getThemeValue } from "../../utils";
+import StyledText from "./StyledText";
+import colors from "../../theme/colors";
 import getRelByTarget from "../../utils/link";
+import { getFontColor } from "../../utils/typography";
 
-const SIZES = {
-  small: {
-    fontSize: typography.size.hecto
-  },
-  large: {
-    fontSize: typography.size.kilo
-  }
-};
+export const LinkTitle = StyledText.withComponent(`a`).extend`
+  color: ${({ color }) => color || colors.azure.base};
+  text-decoration: none;
+`;
 
-const LinkBase = styled.a`
+const LinkBase = Base.withComponent("a").extend`
   text-decoration: none;
   transition: color 0.3s ease;
-  transition: text-decoration 0.3s ease;
-  transition: transform 0.1s linear;
-  font-weight: ${typography.weight.semiBold};
-  font-size: ${({ size }) => SIZES[size].fontSize};
-  line-height: ${typography.lineHeight.header};
-  color: ${getThemeValue("primary", "base")};
-  display: inline-block;
-
-  &:focus {
-    outline: none;
-    text-decoration-line: underline;
-    text-decoration-style: double;
-  }
-
-  &:visited {
-    color: ${getThemeValue("primary", "dark")};
-  }
-
-  &:active {
-    transform: scale(0.98, 0.98) translate(0, 1px);
-    color: ${getThemeValue("primary", "dark")};
+  &:focus,
+  &:active,
+  &:visited,
+  &:hover {
+    color: ${props => getFontColor(props)};
   }
 
   &:hover {
-    color: ${getThemeValue("primary", "medium")};
+    color: ${colors.azure.hover};
   }
 `;
 
-const LinkButtonBase = LinkBase.withComponent(`button`).extend`
+const LinkButtonBase = LinkBase.withComponent("button").extend`
   appearance: none;
   border: 0;
   outline: 0;
@@ -56,7 +37,7 @@ const LinkButtonBase = LinkBase.withComponent(`button`).extend`
   cursor: pointer;
 `;
 
-const LinkSpanBase = LinkBase.withComponent(`span`).extend`
+const LinkSpanBase = LinkBase.withComponent("span").extend`
   cursor: pointer;
 `;
 
@@ -68,18 +49,40 @@ const getElement = ({ href, onClick }) => {
   return LinkSpanBase;
 };
 
-const Link = ({ href, onClick, children, size, ...props }) => {
+const Link = ({
+  href,
+  onClick,
+  children,
+  weight,
+  size,
+  responsiveSize,
+  variant,
+  accent,
+  primary,
+  ...props
+}) => {
   const { target, rel } = props;
   const Elm = getElement({ href, onClick });
   const validatedRel = getRelByTarget(target, rel);
-
   return (
     <Elm
       {...props}
-      size={size}
+      size={{
+        small: responsiveSize.small || size,
+        medium: responsiveSize.medium || responsiveSize.small || size,
+        large:
+          responsiveSize.large ||
+          responsiveSize.medium ||
+          responsiveSize.small ||
+          size
+      }}
+      primary={primary}
+      variant={variant}
+      accent={accent}
       href={href}
-      onClick={onClick}
       rel={validatedRel}
+      weight={weight}
+      onClick={onClick}
     >
       {children}
     </Elm>
@@ -87,18 +90,30 @@ const Link = ({ href, onClick, children, size, ...props }) => {
 };
 
 Link.propTypes = {
-  size: PropTypes.oneOf(Object.values(PT.LINK_SIZES)),
   children: PropTypes.node.isRequired,
   href: PropTypes.string,
-  onClick: PropTypes.func
+  onClick: PropTypes.func,
+  target: PropTypes.string,
+  rel: PropTypes.string,
+  primary: PropTypes.bool,
+  size: PT.size,
+  responsiveSize: PT.responsiveSize,
+  weight: PT.weight,
+  variant: PT.variant,
+  accent: PT.accent
 };
 
 Link.defaultProps = {
-  size: PT.LINK_SIZES.small,
+  target: "",
   onClick: null,
-  href: null
+  href: null,
+  rel: "_self",
+  size: "hecto",
+  responsiveSize: {},
+  weight: "regular",
+  variant: "accent",
+  accent: "azure",
+  primary: true
 };
-
-Link.displayName = "Link";
 
 export default Link;
